Changed from double to int

+++ NavigationSystem/DataBase/DBHandler.h
@@ -154,7 +154,7 @@ public:
 	void deleteRow(std::string table, std::string id);
 
 	bool getWaypointValues(int& nextId, double& nextLongitude, double& nextLatitude, int& nextDeclination, int& nextRadius, int& nextStayTime,
-						double& timeInterval, bool& measureAtCheckpoint,
+						int& timeInterval, bool& measureAtCheckpoint,
                         int& prevId, double& prevLongitude, double& prevLatitude, int& prevDeclination, int& prevRadius, bool& foundPrev);
 
 	bool insert(std::string table, std::string fields, std::string values);
